pyproject.toml
[tool.poetry]
name = "adidascs"
version = "0.1.0"
description = ""
authors = ["Ceyhun Kerti <ceyhun.kerti@bluecolor.io>"]

[tool.poetry.dependencies]
python = "^3.8"
pyspark = "^3.1.1"
python-dotenv = "^0.16.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
